app/requirements.txt
kubernetes==12.0.0
PyYAML==5.4.1
